Fix sky torrents torrent url (#58)
* Cleanup indentation

* Sky Torrents: Fix torrent URL

Currently Sky Torrents RSS gives invalid URL: the path part is not
URL-escaped (it contains spaces, etc). These invalid URLs are not
supported by at least rtorrent: the download is just not done.

java.net.URL constructor accepts such malformed URLs, whereas
java.net.URI does not. So we use URL to parse the URL, and construct a
proper URI, letting URI do the escaping.

We are also future-proof: we first check with URI if the URL is already
valid or not.